requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
decorator==4.1.2
Django==1.11.5
djangorestframework==3.6.4
enum34==1.1.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
